Refactor views: split on several modules
