+++ lupanes/views/customer.py
@@ -1,20 +1,17 @@
-import calendar
-from decimal import Decimal
 from typing import Any, Dict
 
-from django.contrib.auth.mixins import LoginRequiredMixin
 from django.db.models import QuerySet
 from django.http import HttpRequest, HttpResponse, JsonResponse
 from django.urls import reverse_lazy
 from django.utils import timezone
-from django.views.generic import DetailView, ListView, RedirectView
+from django.views.generic import DetailView, RedirectView
 from django.views.generic.edit import (CreateView, DeleteView, FormView,
                                        UpdateView)
 
-from lupanes import helpers
+
 from lupanes.forms import CustomerAuthForm, DeliveryNoteCreateForm
 from lupanes.mixins import CustomerAuthMixin
-from lupanes.models import Customer, DeliveryNote, Product
+from lupanes.models import DeliveryNote, Product
 
 
 class CustomerLoginView(FormView):
@@ -74,58 +71,6 @@ class DeliveryNoteDeleteView(CustomerAuthMixin, DeleteView):
     def get_queryset(self) -> QuerySet[Any]:
         today = timezone.now().date()
         return DeliveryNote.objects.filter(customer=self.customer, date__date=today)
-
-
-class DeliveryNoteListView(LoginRequiredMixin, ListView):
-    model = DeliveryNote
-
-    def get_queryset(self) -> QuerySet[Any]:
-        value = self.request.GET.get("month")
-        self.month = helpers.clean_month(value)
-
-        return DeliveryNote.objects.filter(
-            date__date__year=self.month.year,
-            date__date__month=self.month.month,
-        )
-
-    def get_context_data(self, **kwargs: Any) -> Dict[str, Any]:
-        context = super().get_context_data(**kwargs)
-        today = timezone.now().date()
-        choices = {
-            i: calendar.month_name[i]
-            for i in range(1, today.month + 1)
-        }
-
-        context.update({
-            "month": self.month,
-            "choices": choices,
-        })
-        return context
-
-
-class DeliveryNoteSummaryView(LoginRequiredMixin, ListView):
-    template_name = "lupanes/deliverynote_summary.html"
-
-    def get_queryset(self) -> QuerySet[Any]:
-        value = self.request.GET.get("month")
-        self.period = helpers.clean_month(value)
-
-        qs = Customer.objects.filter(is_active=True)
-        for customer in qs:
-            customer.total = Decimal(0)
-            notes = customer.deliverynote_set.filter(
-                date__date__year=self.period.year,
-                date__date__month=self.period.month,
-            )
-            for note in notes:
-                customer.total += note.amount()
-
-        return qs
-
-    def get_context_data(self, **kwargs: Any) -> Dict[str, Any]:
-        context = super().get_context_data(**kwargs)
-        context["period"] = self.period
-        return context
 
 
 class ProductAjaxView(DetailView):
